Учтены замечания

+++ application/src/main/kotlin/com/unicorn/power/data/Sample.kt
@@ -2,4 +2,4 @@ package com.unicorn.power.data
 
 import java.math.BigDecimal
 
-data class Sample(val sum: Double, val avg: Double, val max: Double, val min: Double, val count: Long)
+data class Sample(val sum: BigDecimal, val avg: BigDecimal, val max: BigDecimal, val min: BigDecimal, val count: Long)
